# Conflicts:
#	src/app/pages/editproject-web/editproject-web.config.js
#	src/app/pages/faq/faq.config.js
#	src/styles/common.scss
#	src/styles/main.scss

+++ src/app/pages/editproject-web/editproject-web.controller.js
@@ -1,13 +1,13 @@
 class EditProjectWebCtrl {
-    constructor(AppConstants, Project, $state, $stateParams, $q, $scope, User, JWT, EventBus, ProjectStore, Notification) {
+    constructor(AppConstants, Project, $state, $stateParams, $q, $scope, User, JWT, EventBus, ProjectStore, Notification, $uibModal) {
         'ngInject';
 
         if (!$stateParams.projectId) return $state.go('landing.error');
 
-
         this.Notification = Notification;
         this.appName = AppConstants.appName;
         this._AppConstants = AppConstants;
+        this._$uibModal = $uibModal;
         this._JWT = JWT;
         this.Project = Project;
         this.$state = $state;
@@ -44,6 +44,7 @@ class EditProjectWebCtrl {
         };
 
         this.submiting = false;
+        this.recordIp = this._AppConstants.RECORDIP;
         this.openEvent = null;
         this.domainPattern = /^(?:(?:https?|ftp):\/\/)?(?:\S+(?::\S*)?@)?(?:(?!(?:10|127)(?:\.\d{1,3}){3})(?!(?:169\.254|192\.168)(?:\.\d{1,3}){2})(?!172\.(?:1[6-9]|2\d|3[0-1])(?:\.\d{1,3}){2})(?:[1-9]\d?|1\d\d|2[01]\d|22[0-3])(?:\.(?:1?\d{1,2}|2[0-4]\d|25[0-5])){2}(?:\.(?:[1-9]\d?|1\d\d|2[0-4]\d|25[0-4]))|(?:(?:[a-z\u00a1-\uffff0-9]-*)*[a-z\u00a1-\uffff0-9]+)(?:\.(?:[a-z\u00a1-\uffff0-9]-*)*[a-z\u00a1-\uffff0-9]+)*(?:\.(?:[a-z\u00a1-\uffff]{2,}))\.?)(?::\d{2,5})?(?:[/?#]\S*)?$/i;
         this.domain = '';
@@ -55,6 +56,8 @@ class EditProjectWebCtrl {
             else
                 this.finalizeRequest();
         });
+
+
     }
 
     getProject() {
@@ -79,35 +82,87 @@ class EditProjectWebCtrl {
     }
 
     addDomain() {
-        this.domain = this.domain.replace(/.*?:\/\//g, '');
-        this.Project.setDomain({ id: this.projectId, domain: this.domain }).then(
+
+        if (this.project.domain && this.project.domain !== this.domain) {
+            this.modalText = `<span class="text-highlight">${this.project.domain}</span> existing domain parameters will be overwritten to <span class="text-highlight">${this.domain}</span>`;
+            this.delete = false;
+            return this.modalInstance = this._$uibModal.open({
+                animation: this.animationsEnabled,
+                ariaLabelledBy: 'modal-title',
+                ariaDescribedBy: 'modal-body',
+                templateUrl: 'layout/modals/domainExists.html',
+                //controller: EditProjectWebCtrl,
+                controllerAs: 'vm',
+                bindToController: true,
+                scope: this._$scope,
+                resolve: {},
+            });
+        }
+        this.submitDomain();
+    }
+
+    submitDomain() {
+        if(this.domainAddInProgress) return;
+        this.domainAddInProgress = true;
+        const domain = angular.copy(this.domain).replace(/.*?:\/\//g, '');
+        this.Project.setDomain({ id: this.projectId, domain: domain }).then(
             response => {
                 this.Notification.success(response.message);
-                this.project.domain = this.domain;
+                this.project.domain = domain;
+                this.domain = domain;
+                this.domainAddInProgress = false;
+                if (this.modalInstance) this.modalInstance.close();
             },
 
             err=> {
                 _.each(err, (val, key)=> {
                     this.Notification.error(val.fieldName);
                 });
-                console.log(err);
+                this.domainAddInProgress = false;
+                if (this.modalInstance) this.modalInstance.close();
             }
         );
     }
 
     deleteDomain() {
+        this.modalText = `Are you sure you want delete <span class="text-highlight">${this.project.domain}</span> ?`;
+        this.delete = true;
+        return this.modalInstance = this._$uibModal.open({
+            animation: this.animationsEnabled,
+            ariaLabelledBy: 'modal-title',
+            ariaDescribedBy: 'modal-body',
+            templateUrl: 'layout/modals/domainExists.html',
+            //controller: EditProjectWebCtrl,
+            controllerAs: 'vm',
+            bindToController: true,
+            scope: this._$scope,
+            resolve: {},
+        });
+    }
+
+    comfirmDelete() {
+        if(this.deleteInProgress) return;
+        this.deleteInProgress = true;
         this.Project.deleteDomain(this.projectId, this.project).then(
             response => {
                 this.Notification.success(response.message);
+                this.deleteInProgress = false;
+                if (this.modalInstance) this.modalInstance.close();
             },
 
             err=> {
                 _.each(err, (val, key)=> {
                     this.Notification.error(val.fieldName);
                 });
-                console.log(err);
+                this.deleteInProgress = false;
+                if (this.modalInstance) this.modalInstance.close();
             }
         );
+    }
+
+    closeModal(){
+        this.domain = this.project.domain;
+        this.modalInstance.close();
     }
 
     switchDomainTrigger() {
